backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.1.0
django-environ==0.10.0
django-extensions==3.2.3
djangorestframework==3.14.0
exceptiongroup==1.1.2
fastapi==0.100.0
idna==3.4
psycopg2-binary==2.9.6
pydantic==2.0.2
pydantic_core==2.1.2
pytz==2023.3
sniffio==1.3.0
SQLAlchemy==2.0.18
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.7.1
